Update writewhatwhere.py
get PsISP returned a pointer to a pointer, not the pointer we wanted. I then used the kernel write (not the read!) to leak that address pointed to by the pointer back to userland and read it. I discovered this bug (the lack of a pointer to a pointer) by reviewing Gradius' code and realizing I was missing a read statement and that that read was checking the value of the pointer PsISP. So, then I went in to the debugger to make sure my addresses leaked prior to that were correct, and they both were and weren't. I was clearly returning a legitimate value, just not the value of the pointer we wanted. We had a pointer to the value we wanted (another pointer). When checking in Windbg I dumped the memory and found the PsISP pointer was pointing to a second pointer, and then used the kernel write to write that value and echo that back to the user. 

So now the code works up past the get SYSTEM process base method.